requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
openai==2.7.1
python-dotenv==1.1.1
discord.py==2.3.2
apscheduler==3.10.4
pytz==2023.3.post1
